requirements.txt
apache-airflow>=2.7.0
apache-airflow-providers-mysql>=5.0.0
apache-airflow-providers-postgres>=5.0.0
kaggle>=1.5.16
pandas>=2.0.0
matplotlib>=3.7.0
sqlalchemy>=1.4.0,<2.0.0
pymysql>=1.1.0
python-dotenv>=0.19.0 
streamlit>=1.20.0
plotly>=5.18.0